pyproject.toml
[project]
name = "database-migration-test-tool"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyodbc",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
]
